(feat) Direct add_rider and remove_rider action types to riders reducer and create handlers for those action types.
