+++ client/reducers/ride.js
@@ -22,8 +22,10 @@ export default function(state = Map(), action) {
     case 'CANCEL_RIDE':
     case 'CANCEL_RIDE_SENT':
     case 'CANCEL_RIDE_ERROR':
-    case 'REMOVE_RIDER':
       return handleCancel(state, action);
+    case 'REMOVE_RIDER':
+    case 'ADD_RIDER':
+      return handleRiders(state.get('riders'), action);
     case 'REQUEST_MESSAGES':
     case 'RECEIVE_MESSAGES':
     case 'REQUEST_MESSAGES_ERROR':
